+++ Woof/Woof/View/OwnerFlow/OwnerMainTabView/OwnerMainTabView.swift
@@ -5,7 +5,7 @@ import SwiftUI
  This view displays a tab bar with different tabs for managing pet sitters, walkings, and owner profile information.
  */
 struct OwnerMainTabView: View {
-    @State private var selection: Tab = .sitters
+    // MARK: - Internal interface
 
     init() {
         customizeTabBar()
@@ -37,8 +37,45 @@ struct OwnerMainTabView: View {
             .tint(Color.App.purpleDark)
             .navigationTitle(selection.header)
             .navigationBarTitleDisplayMode(.inline)
+            .toolbar {
+                NavigationLink(
+                    destination: LoginView()
+                        .navigationBarBackButtonHidden(true),
+                    isActive: $viewModel.isLogoutConfirmed
+                ) {
+                    Button(logoutButtonLabelText) {
+                        viewModel.isAlertShown.toggle()
+                    }
+                }
+            }
+            .foregroundColor(.App.purpleDark)
+
+            .alert(alertTitle, isPresented: $viewModel.isAlertShown) {
+                Button(continueButtonLabelText) {
+                    viewModel.isLogoutConfirmed.toggle()
+                    userRoleViewModel.resetCurrentRole()
+                    dismiss()
+                }
+                Button(
+                    cancelButtonLabelText,
+                    role: .cancel
+                ) { viewModel.isAlertShown.toggle() }
+            }
         }
     }
+
+    // MARK: - Private interface
+
+    @StateObject private var viewModel = OwnerMainTabViewModel()
+    @State private var selection: Tab = .sitters
+
+    @EnvironmentObject private var userRoleViewModel: UserRoleViewModel
+    @Environment(\.dismiss) private var dismiss
+
+    private let logoutButtonLabelText = "Logout"
+    private let continueButtonLabelText = "Continue"
+    private let cancelButtonLabelText = "Cancel"
+    private let alertTitle = "Do you really want to log out?"
 
     /**
      Customizes the appearance of the tab bar.