+++ Woof/WoofTests/ViewModel/SitterMainTabViewModelTests.swift
@@ -1,18 +1,56 @@
 import XCTest
 
 final class SitterMainTabViewModelTests: XCTestCase {
-    func testResetCurrentRoleMethodExistsInAPI() {
-        SitterMainTabViewModel().resetCurrentRole()
+    override func setUp() {
+        super.setUp()
+        viewModel = SitterMainTabViewModel()
     }
 
-    func testResetCurrentRoleChangesUserRoleToNone() {
-        // Given
-        PreferencesHandler.set(userRole: .sitter)
+    func testIsAlertShownPropertyExistsInAPI() {
+        _ = viewModel.isAlertShown
+    }
 
-        // When
-        SitterMainTabViewModel().resetCurrentRole()
+    func testIsAlertShownPropertyReturnsAssignedValue() {
+        // Given // When
+        viewModel.isAlertShown = true
 
         // Then
-        XCTAssertEqual(PreferencesHandler.getUserRole(), Role.none)
+        XCTAssertTrue(viewModel.isAlertShown)
+
+        // Given // When
+        viewModel.isAlertShown = false
+
+        // Then
+        XCTAssertFalse(viewModel.isAlertShown)
     }
+
+    func testIsAlertShownPropertyReturnsFalseByDefault() {
+        XCTAssertFalse(viewModel.isAlertShown)
+    }
+
+    func testIsLogoutConfirmedPropertyExistsInAPI() {
+        _ = viewModel.isLogoutConfirmed
+    }
+
+    func testIsLogoutConfirmedPropertyReturnsAssignedValue() {
+        // Given // When
+        viewModel.isLogoutConfirmed = true
+
+        // Then
+        XCTAssertTrue(viewModel.isLogoutConfirmed)
+
+        // Given // When
+        viewModel.isLogoutConfirmed = false
+
+        // Then
+        XCTAssertFalse(viewModel.isLogoutConfirmed)
+    }
+
+    func testIsLogoutConfirmedPropertyReturnsFalseByDefault() {
+        XCTAssertFalse(viewModel.isLogoutConfirmed)
+    }
+
+    // MARK: - Private interface
+
+    private var viewModel = SitterMainTabViewModel()
 }